requirements.txt
fastapi==0.115.11
uvicorn==0.23.2
langgraph==0.0.10
sqlalchemy==2.0.0
faiss-cpu==1.7.4
openai==1.0.0
pydantic==2.0.0
python-dotenv==1.0.0
httpx==0.27.0
jinja2==3.1.4
alembic==1.15.2
numpy==1.24.0
GitPython==3.1.43
langchain-core==0.1.0
langchain-openai==0.0.1 